# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Cloud Gaming signaling, session, and wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/{userToken}:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Wallet overview — balance, deposit addresses, withdrawal history
      parameters:
        - in: path
          name: userToken
          required: true
          schema:
            type: string
          description: Either a host token or a player token
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/{userToken}/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallet]
      summary: Recent wallet transactions (deposits, withdrawals, billing)
      parameters:
        - in: path
          name: userToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/{userToken}/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Submit a withdrawal request (queued for processing)
      parameters:
        - in: path
          name: userToken
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestWithdrawalBody"
      responses:
        "201":
          description: Withdrawal queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Invalid request (e.g. insufficient balance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DepositAddress:
      type: object
      properties:
        currency:
          type: string
          description: "USDT_TRC20 | NANO | SOL"
        label:
          type: string
        address:
          type: string
        network:
          type: string
        minDeposit:
          type: number
      required:
        - currency
        - label
        - address
        - network
        - minDeposit
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        ownerType:
          type: string
        ownerId:
          type: string
        currency:
          type: string
        address:
          type: string
        amountUsdt:
          type: number
          description: Amount in crypto USDT-equivalent (LZT / 200)
        amountLzt:
          type: integer
        status:
          type: string
        requestedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - ownerType
        - ownerId
        - currency
        - address
        - amountUsdt
        - amountLzt
        - status
        - requestedAt
    RequestWithdrawalBody:
      type: object
      properties:
        currency:
          type: string
        address:
          type: string
        amountLzt:
          type: integer
          description: Amount in LZT (must be a multiple of 200 — i.e. whole USDT)
      required:
        - currency
        - address
        - amountLzt
    Wallet:
      type: object
      properties:
        ownerType:
          type: string
          description: "host | player"
        ownerId:
          type: string
        displayName:
          type: string
        internalBalanceLzt:
          type: integer
        withdrawableBalanceLzt:
          type: integer
        pendingWithdrawalsLzt:
          type: integer
        lztPerUsdt:
          type: integer
        depositAddresses:
          type: array
          items:
            $ref: "#/components/schemas/DepositAddress"
        recentWithdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
      required:
        - ownerType
        - ownerId
        - displayName
        - internalBalanceLzt
        - withdrawableBalanceLzt
        - pendingWithdrawalsLzt
        - lztPerUsdt
        - depositAddresses
        - recentWithdrawals
...
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          description: "deposit | withdrawal | session_billing"
        currency:
          type: ["string", "null"]
        amountLzt:
          type: integer
          description: Signed LZT (negative for debits, positive for credits)
        status:
          type: ["string", "null"]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - kind
        - amountLzt
        - description
        - timestamp