# lib/api-zod/src/generated/types/requestWithdrawalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestWithdrawalBody {
  currency: string;
  address: string;
  /** Amount in LZT (must be a multiple of 200 — i.e. whole USDT) */
  amountLzt: number;
}