# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { DepositAddress } from "./depositAddress";
import type { Withdrawal } from "./withdrawal";

export interface Wallet {
  /** host | player */
  ownerType: string;
  ownerId: string;
  displayName: string;
  internalBalanceLzt: number;
  withdrawableBalanceLzt: number;
  pendingWithdrawalsLzt: number;
  lztPerUsdt: number;
  depositAddresses: DepositAddress[];
  recentWithdrawals: Withdrawal[];
}