# lib/api-zod/src/generated/types/walletTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletTransaction {
  id: string;
  /** deposit | withdrawal | session_billing */
  kind: string;
  /** @nullable */
  currency?: string | null;
  /** Signed LZT (negative for debits, positive for credits) */
  amountLzt: number;
  /** @nullable */
  status?: string | null;
  description: string;
  timestamp: Date;
}